requirements.txt
Flask
tweepy
PyYAML
requests
setuptools
Unidecode